// src/auth/dto/user.dto.ts
export class userDetailSendToClient {
  CUSTOMER_ID: string;
  GENDER: string;
  FIRST_NAME: string;
  LAST_NAME: string;
  EMAIL_ADDRESS: string;
  LOGIN_NAME: string;
  LOGIN_PASSWORD: string;
  PHONE_NUMBER: string;
  ADDRESS: string;
}
export class userSummaryInfo {
  CUSTOMER_ID: Number;
  FIRST_NAME: string;
  AVATAR: string;
  ROLE?: string;
}